# Separate `BaseProvider` for faster import