--- a/packages/jupyter-ai-magics/jupyter_ai_magics/__init__.py
+++ b/packages/jupyter-ai-magics/jupyter_ai_magics/__init__.py
@@ -1,42 +1,63 @@
+from typing import TYPE_CHECKING
+
 from ._import_utils import import_attr
 from ._version import __version__
 
-_dynamic_imports = {
-    # expose embedding model providers on the package root
-    "embedding_providers": [
-        "BaseEmbeddingsProvider",
-        "GPT4AllEmbeddingsProvider",
-        "HfHubEmbeddingsProvider",
-        "QianfanEmbeddingsEndpointProvider",
-    ],
-    "exception": ["store_exception"],
-    "magics": ["AiMagics"],
-    # expose JupyternautPersona on the package root
-    # required by `jupyter-ai`.
-    "models.persona": ["JupyternautPersona", "Persona"],
-    # expose model providers on the package root
-    "providers": [
-        "AI21Provider",
-        "BaseProvider",
-        "GPT4AllProvider",
-        "HfHubProvider",
-        "QianfanProvider",
-        "TogetherAIProvider",
-    ],
-}
+if TYPE_CHECKING:
+    # same as dynamic imports but understood by mypy
+    from .embedding_providers import (
+        BaseEmbeddingsProvider,
+        GPT4AllEmbeddingsProvider,
+        HfHubEmbeddingsProvider,
+        QianfanEmbeddingsEndpointProvider,
+    )
+    from .exception import store_exception
+    from .magics import AiMagics
+    from .models.persona import JupyternautPersona, Persona
+    from .providers import (
+        AI21Provider,
+        BaseProvider,
+        GPT4AllProvider,
+        HfHubProvider,
+        QianfanProvider,
+        TogetherAIProvider,
+    )
+else:
+    _dynamic_imports = {
+        # expose embedding model providers on the package root
+        "embedding_providers": [
+            "BaseEmbeddingsProvider",
+            "GPT4AllEmbeddingsProvider",
+            "HfHubEmbeddingsProvider",
+            "QianfanEmbeddingsEndpointProvider",
+        ],
+        "exception": ["store_exception"],
+        "magics": ["AiMagics"],
+        # expose JupyternautPersona on the package root
+        # required by `jupyter-ai`.
+        "models.persona": ["JupyternautPersona", "Persona"],
+        # expose model providers on the package root
+        "providers": [
+            "AI21Provider",
+            "BaseProvider",
+            "GPT4AllProvider",
+            "HfHubProvider",
+            "QianfanProvider",
+            "TogetherAIProvider",
+        ],
+    }
 
-_dynamic_imports_map = {
-    import_name: module
-    for module, imports in _dynamic_imports.items()
-    for import_name in imports
-}
+    _dynamic_imports_map = {
+        import_name: module
+        for module, imports in _dynamic_imports.items()
+        for import_name in imports
+    }
 
-
-def __getattr__(attr_name: str) -> object:
-    module_name = _dynamic_imports_map.get(attr_name)
-    result = import_attr(attr_name, module_name, __spec__.parent)
-    globals()[attr_name] = result
-    return result
+    def __getattr__(attr_name: str) -> object:
+        module_name = _dynamic_imports_map.get(attr_name)
+        result = import_attr(attr_name, module_name, __spec__.parent)
+        globals()[attr_name] = result
+        return result
 
 
 def load_ipython_extension(ipython):
@@ -48,7 +69,7 @@ def unload_ipython_extension(ipython):
     ipython.set_custom_exc((BaseException,), ipython.CustomTB)
 
 
-# required to preserve backward compatibiliy with `from jupyter_ai_magics import *`
+# required to preserve backward compatibility with `from jupyter_ai_magics import *`
 __all__ = [
     "__version__",
     "load_ipython_extension",
@@ -68,4 +89,8 @@ __all__ = [
     "QianfanProvider",
     "TogetherAIProvider",
 ]
-assert set(_dynamic_imports_map.keys()) - set(__all__) == set()
+
+
+def __dir__():
+    # Allows more editors (e.g. IPython) to complete on `jupyter_ai_magics.<tab>`
+    return list(__all__)