--- a/packages/jupyter-ai-magics/jupyter_ai_magics/partner_providers/anthropic.py
+++ b/packages/jupyter-ai-magics/jupyter_ai_magics/partner_providers/anthropic.py
@@ -1,6 +1,6 @@
 from langchain_anthropic import ChatAnthropic
 
-from ..providers import BaseProvider, EnvAuthStrategy
+from ..base_provider import BaseProvider, EnvAuthStrategy
 
 
 class ChatAnthropicProvider(