--- a/packages/jupyter-ai-magics/jupyter_ai_magics/partner_providers/mistralai.py
+++ b/packages/jupyter-ai-magics/jupyter_ai_magics/partner_providers/mistralai.py
@@ -1,4 +1,4 @@
-from jupyter_ai_magics.providers import BaseProvider, EnvAuthStrategy
+from jupyter_ai_magics.base_provider import BaseProvider, EnvAuthStrategy
 from langchain_mistralai import ChatMistralAI, MistralAIEmbeddings
 
 from ..embedding_providers import BaseEmbeddingsProvider