--- a/packages/jupyter-ai-magics/jupyter_ai_magics/partner_providers/ollama.py
+++ b/packages/jupyter-ai-magics/jupyter_ai_magics/partner_providers/ollama.py
@@ -1,7 +1,7 @@
 from langchain_ollama import ChatOllama, OllamaEmbeddings
 
+from ..base_provider import BaseProvider, TextField
 from ..embedding_providers import BaseEmbeddingsProvider
-from ..providers import BaseProvider, TextField
 
 
 class OllamaProvider(BaseProvider, ChatOllama):