--- a/packages/jupyter-ai-magics/jupyter_ai_magics/partner_providers/openrouter.py
+++ b/packages/jupyter-ai-magics/jupyter_ai_magics/partner_providers/openrouter.py
@@ -1,5 +1,5 @@
 from jupyter_ai_magics import BaseProvider
-from jupyter_ai_magics.providers import EnvAuthStrategy, TextField
+from jupyter_ai_magics.base_provider import EnvAuthStrategy, TextField
 from langchain_core.utils import get_from_dict_or_env
 from langchain_openai import ChatOpenAI