… packages/jupyter-ai-magics/jupyter_ai_magics/base_provider.py
… packages/jupyter-ai-magics/jupyter_ai_magics/providers.py