### References

- Closes #1337

### User-facing changes

IPython users who define a custom model by inheriting from `BaseProvider` will be able to start IPython about 33% faster; this is significant both in relative and absolute terms (while for me it is only 0.4 faster, I run the benchmarks on a very fast CPU which is not necessarily representative of the general user base).

| Import | Before | After |
|--|--|--|
| `from jupyter_ai_magics import BaseProvider` | 1.342 s ±  0.021 s | 1.046 s ±  0.014 s |
| `from jupyter_ai_magics.providers import BaseProvider` | 1.344 s ±  0.016 s | 1.049 s ±  0.010 s | 
| `from jupyter_ai_magics.base_provider import BaseProvider` | (not possible) | 902.5 ms ±  10.0 ms |

<details>

#### Before
```
$ hyperfine "python -c 'from jupyter_ai_magics import BaseProvider'"
Benchmark 1: python -c 'from jupyter_ai_magics import BaseProvider'
  Time (mean ± σ):      1.342 s ±  0.021 s    [User: 1.188 s, System: 0.157 s]
  Range (min … max):    1.315 s …  1.388 s    10 runs
$ hyperfine "python -c 'from jupyter_ai_magics.providers import BaseProvider'"
Benchmark 1: python -c 'from jupyter_ai_magics.providers import BaseProvider'
  Time (mean ± σ):      1.344 s ±  0.016 s    [User: 1.194 s, System: 0.153 s]
  Range (min … max):    1.318 s …  1.368 s    10 runs

```

#### After

```
$ hyperfine "python -c 'from jupyter_ai_magics import BaseProvider'"
Benchmark 1: python -c 'from jupyter_ai_magics import BaseProvider'
  Time (mean ± σ):      1.046 s ±  0.014 s    [User: 0.914 s, System: 0.136 s]
  Range (min … max):    1.022 s …  1.072 s    10 runs
$ hyperfine "python -c 'from jupyter_ai_magics.providers import BaseProvider'"
Benchmark 1: python -c 'from jupyter_ai_magics.providers import BaseProvider'
  Time (mean ± σ):      1.049 s ±  0.010 s    [User: 0.924 s, System: 0.129 s]
  Range (min … max):    1.028 s …  1.058 s    10 runs
$ hyperfine "python -c 'from jupyter_ai_magics.base_provider import BaseProvider'"
Benchmark 1: python -c 'from jupyter_ai_magics.base_provider import BaseProvider'
  Time (mean ± σ):     902.5 ms ±  10.0 ms    [User: 788.1 ms, System: 118.4 ms]
  Range (min … max):   887.7 ms … 914.9 ms    10 runs
```

</details>

`jupyter-ai-magics` users who use `%load_ext jupyter_ai_magics` at startup will not see a speedup yet because `self.providers = get_lm_providers()` in `AiMagics` init causes all providers to be loaded. They would only see a speedup if they disabled some of the providers (which previously would not have been the case as all providers would be imported anyways).

### To be done

- [x] vendor `import_attr` instead of importing it from `langchain_core._import_utils` to make sure it does not break in a point release of `langchain` (preserving the attribution to langchain team)
- [x] <s>fix typing</s> actually, `mypy` is failing on other PRs too..