@dlqqq 2025-05-13
@krassowski Thank you for opening this PR! Really appreciate you taking the lead on debugging the module import latency & for providing objective data. I remember you had mentioned in another thread that LangChain recently improved their module imports to be faster; do you know if that is as simple as upgrading our version floor on `langchain`? Would be curious to see if that makes an impact too.

I've been tasked with a 2-week coding sprint to work on another project right now. Sorry if I'm slow to respond for the next week. I'll be back on Monday to give this a full review.