@krassowski 2025-05-13
> I remember you had mentioned in another thread that LangChain recently improved their module imports to be faster; do you know if that is as simple as upgrading our version floor on langchain?

Yup, this was langchain-core, and the timings in this PR are based on the latest version including the fix. In fact this PR in the current state directly imports the lazy import machinery from `langchain_core` (which we should not do as it is private, we should vendor a copy instead), because this approach is required to ensure the imports are not triggered in `__init__.py`.

Separately of this PR if you want to ensure all users get the faster imports then yes adding floor level pin should be enough as it was released in a minor version.

Have a good coding spring! I saw your status which is why I did not ping - I really appreciate a quick look from you!