@dlqqq 2025-05-13
BTW, do you have a good reference on how Python module imports work? I'm surprised that you & LangChain maintainers had to add manual code changes to get dynamic fast imports.