@krassowski 2025-05-13
The relevant bits are:
- https://peps.python.org/pep-0562/ looking at it we may want to implement `__dir__` too 
- https://docs.python.org/3/reference/import.html#regular-packages