@krassowski 2025-05-30
@dlqqq I implemented the suggested changes and can now answer your questions:

> When consumers import from `jupyter_ai_magics`, do they still get type inference for mypy
Yes!

While `mypy` supports module-level `__getattr__` for a long time now (https://github.com/python/mypy/pull/6606), it would require manually adding `@overload`s for each import.

A simpler approach is to just use `typing.TYPE_CHECKING` which I have now done.  You can verify that this now works by looking at CI which is now passing but was failing prior to https://github.com/jupyterlab/jupyter-ai/pull/1338/commits/d8adcbb0a17349b8eec6fcdfd44523d8da3ad965. I also added a test to verify that the modules include in the `TYPE_CHECKING` block are the same as the one being dynamically imported in https://github.com/jupyterlab/jupyter-ai/pull/1338/commits/65774361cf47d07f3eafa037f8f73e5a9b41dac1.


> & IDE intellisense?

Yes! For IPython and therefore Jupyter this now works after adding `__dir__` implementation:

![image](https://github.com/user-attachments/assets/54c8dfd6-e147-45c8-8bcf-39480d4d7f1c)

It also works with the alternative completer implementation:

![image](https://github.com/user-attachments/assets/76782c88-9353-4302-be7a-0ce4af7d016e)

It also works fine in VSCode, although it includes the `import_attr` which should not be shown as per out `__dir__` and `__all__`:

![image](https://github.com/user-attachments/assets/6a1a5596-33a2-482a-b4bc-7f2972496c17)

We could rename `import_attr` to `_import_attr`... I will do it right now...