@dlqqq 2025-06-02
BTW, in this PR, server extension startup time now logs ~1450ms instead of ~1600ms on my machine. Even though it's slight, I noticed it as a user, which is why I compared the logs. 🤗