@krassowski 2025-06-03
> BTW, in this PR, server extension startup time now logs ~1450ms instead of ~1600ms on my machine

Right, I did not notice earlier, but indeed it does improve the startup time for jupyterlab users too! 

`hyperfine 'jupyter server extension list'`

| Before | After | Speedup |
|--|--|--|
| 2.683 s ±  0.021 s | 2.413 s ±  0.021 s | 10.0% |

`hyperfine "python -c 'import jupyter_ai'"`

| Before | After | Speedup |
|--|--|--|
| 2.289 s ±  0.039 s | 2.016 s ±  0.052 s | 11.9% |