@krassowski 2025-06-03
In a follow-up PR we can explore changing imports in `jupyter-ai` itself. In particular this:

https://github.com/jupyterlab/jupyter-ai/blob/9e535fde92c0de3f2f1743c3de7923e255030c78/packages/jupyter-ai/jupyter_ai/__init__.py#L10-L12

appears to be worth making into a lazy import, because it forces loading of all provider classes, even though none are needed in the extension use case (it is basically a convenience to allow to import from `jupyter_ai`)