@krassowski 2025-06-03
@meeseeksdev please backport to 2.x